Add a "cutoff" distance to the distance calculator.  This value can be used to short circuit having to calculate the distance between every pair in the candidate cluster because if the distance of any single pair is greater than the diameter of another candidate, we know it's already too big.
Note I also changed a comparison in the qtclusterer from "<=" to "<".  The reason for the change is simply to avoid replacing a perfectly good candidate cluster with another  that has the exact same diameter.  The side effect of this may change the "order" of instances in a particular clusters generated with older code, but the end result is equivalent.   For this reason, some unit tests that check exact order of the results have been updated.

git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/trunk/icb-commons/qtclustering@14351 60cd2ccc-2b05-0410-9566-da273a940bff